backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Werkzeug
requests
PyJWT
